+++ main.c
@@ -7,9 +7,9 @@
 
 
 
-void print_1d_arr(long double * arr, int N){
+void print_1d_arr(double * arr, int N){
     for (int i = 0; i < N; i++){
-        printf("%Lf ", arr[i]);
+        printf("%f ", arr[i]);
     }
     printf("\n");
 }
@@ -29,11 +29,11 @@ int get_n(int argc, char* argv[]){
     }
     return N;
 }
-void read_input(long double ** basis, int argc, char* argv[], int N){
+void read_input(double ** basis, int argc, char* argv[], int N){
     int index = 0;
     int extra = 0;
     for (int i = 0; i < N; i++){
-        basis[i] = malloc(sizeof(long double)*N);
+        basis[i] = malloc(sizeof(double)*N);
         for (int j = 0; j<N; j++){
             if (argv[index][0] == '['){
                 extra = 1;
@@ -47,10 +47,10 @@ void read_input(long double ** basis, int argc, char* argv[], int N){
     }
 
 }
-void print_2d_arr(long double ** arr, int N){
+void print_2d_arr(double ** arr, int N){
     for (int i = 0; i < N; i++){
         for (int j = 0; j < N; j++){
-            printf("%Lf ", arr[i][j]);
+            printf("%f ", arr[i][j]);
         }
         printf("\n");
     }
@@ -58,14 +58,14 @@ void print_2d_arr(long double ** arr, int N){
 
 int main(int argc, char *argv[]) {
     int N = get_n(argc-1, argv+1);
-    long double ** basis = malloc(N*sizeof(long double));
+    double ** basis = malloc(N*sizeof(double));
     read_input(basis, argc-1, argv+1, N);
     // print_2d_arr(basis, N);
-    long double ans;
+    double ans;
 
     ans = svp(basis, N);
     // for (int i = 0; i < N; i++){ 
-    //     printf("%Lf\n", sqrt(dot_product(basis[i], basis[i], N)));
+    //     printf("%f\n", sqrt(dot_product(basis[i], basis[i], N)));
     // }
     FILE *file = fopen("result.txt", "w");
 
@@ -74,7 +74,7 @@ int main(int argc, char *argv[]) {
         return -1;
     }
 
-    fprintf(file, "%Lf", ans);
+    fprintf(file, "%f", ans);
     fclose(file);
     
     return 0;
